split Version and PipeToFile
